mcp-phi-base/workers/discovery/requirements.txt
# Discovery Worker Dependencies
numpy==1.24.3
scipy==1.10.1
sympy==1.12
networkx==3.1
pika==1.3.2  # RabbitMQ
redis==4.5.5
psycopg2-binary==2.9.6
python-dotenv==1.0.0
structlog==23.1.0  # Structured logging